konfigurasi multer (uploads) electshop

+++ src/controllers/shopController.js
@@ -100,6 +100,10 @@ const getAllShop = async (req, res) => {
 const removeShopById = async (req, res) => {
     try {
         const { shop_id } = req.params
+
+        const equalShopId = await shopModel.findOne({shop_id})
+        if(!equalShopId) return res.json({ status: 404, message: 'Shop not found!' })
+        
         const dataShopDelete = await shopModel.deleteOne({shop_id})
 
         if(!dataShopDelete) return res.json({ status: 404, message: 'Shop not found!' })
